// frontend/src/app/generated/model/v1Workflow.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { V1WorkflowStatus } from './v1WorkflowStatus';
import { V1WorkflowMeta } from './v1WorkflowMeta';

export interface V1Workflow {
  generation?: number;
  id?: string;
  meta?: V1WorkflowMeta;
  resourceVersion?: number;
  status?: V1WorkflowStatus;
}