// frontend/src/app/generated/model/v1TaskInstance.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { V1TaskInstanceMeta } from './v1TaskInstanceMeta';
import { V1TaskInstanceStatus } from './v1TaskInstanceStatus';

export interface V1TaskInstance {
  generation?: number;
  id?: string;
  meta?: V1TaskInstanceMeta;
  resourceVersion?: number;
  status?: V1TaskInstanceStatus;
}